Migrate subscript query building to scalar
Relates to https://github.com/crate/crate/issues/10025
